对Bestman_sim中接口进行整理，删除了不必要的接口
